Add Some Engine Features
Engine - Manages the overall functionality of the game program. It also instantiates and loads managers that are used throughout the program, making them accessible globally within the game.

Timer - Manages time in the game. You can specify a frame at the start of the program.

Window - Windows in the game program. When you create it, you can specify the size and name of the window as you want.

GameStateManager - Manages game states, including functions such as level creation, updating, and changing.

Test levels - You can move to other level by pressing keys 1 and 2.
1 - Shader Test
2 - Vertex Test
(If you repeatedly remove and reload vertices and textures, memory continues to accumulate in levels. If there is a correct method about memory management, please advise on it.)

- NOTICE - This is not FINAL VERSION